+++ src/solstice/global.lua
@@ -145,8 +145,12 @@ function _SOL_GET_CACHED_OBJECT(id)
       if object.id == id then
          object.load_char_finished = true
       end
-
-      object.ci   = C.Local_GetCombatInfo(id, object.id ~= id)
+      local ci = Game.GetProperty(id, "SOL_COMBAT_INFO")
+      if not ci then
+         ci = combat_info_t()
+         Game.SetProperty(id, "SOL_COMBAT_INFO", ci)
+      end
+      object.ci = ci
       assert(object.ci ~= nil, "CombatInfo cannot be nil...")
    elseif type == OBJECT_TRUETYPE_MODULE then
       object.obj = C.nwn_GetModule()